Version 1/Python-Client/requirements.txt
MouseInfo==0.1.3
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pyserial==3.5
PyTweening==1.0.3
rubicon-objc==0.4.0
